Refactor schema builder into a class in OData handler
